+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ and by implication, [Semantic Versioning](http://semver.org/).
 - Updated get_checklist to return the number of observers.
 - Updated get_checklist to use table for extracting protocol values.
 - Updated get_checklist to return duration in minutes.
+- Updated get_checklist to return distance in metres.
 
 [0.2.0] - 2024-10-23
 --------------------